Task-50933 : Fix feedbacks

+++ services/src/main/java/org/exoplatform/news/service/impl/NewsServiceImpl.java
@@ -145,7 +145,7 @@ public class NewsServiceImpl implements NewsService {
       }
     }
     List<String> oldTargets = newsTargetingService.getTargetsByNewsId(news.getId());
-    if(publish == news.isPublished() && news.isPublished() && news.isCanPublish() && !oldTargets.equals(news.getTargets())) {
+    if(publish == news.isPublished() && news.isPublished() && news.isCanPublish() && news.getTargets() != null && !oldTargets.equals(news.getTargets())) {
       newsTargetingService.deleteNewsTargets(news.getId());
       newsTargetingService.saveNewsTarget(news.getId(), news.getTargets(), updater);
     }
@@ -304,9 +304,9 @@ public class NewsServiceImpl implements NewsService {
       throw new IllegalArgumentException("User " + currentIdentity.getUserId() + " is not authorized to schedule news");
     }
     List<String> oldTargets = newsTargetingService.getTargetsByNewsId(news.getId());
-    if(!oldTargets.equals(news.getTargets())) {
+    if(news.getTargets() != null && !oldTargets.equals(news.getTargets())) {
       newsTargetingService.deleteNewsTargets(news.getId());
-        newsTargetingService.saveNewsTarget(news.getId(), news.getTargets(), currentIdentity.getUserId());
+      newsTargetingService.saveNewsTarget(news.getId(), news.getTargets(), currentIdentity.getUserId());
     }
     return newsStorage.scheduleNews(news);
   }
@@ -366,9 +366,9 @@ public class NewsServiceImpl implements NewsService {
     News news = getNewsById(newNews.getId(), false);
     newsStorage.publishNews(news);
     List<String> oldTargets = newsTargetingService.getTargetsByNewsId(newNews.getId());
-    if(!oldTargets.equals(newNews.getTargets())) {
+    if(newNews.getTargets() != null && !oldTargets.equals(newNews.getTargets())) {
       newsTargetingService.deleteNewsTargets(newNews.getId());
-        newsTargetingService.saveNewsTarget(newNews.getId(), newNews.getTargets(), publisher);
+      newsTargetingService.saveNewsTarget(newNews.getId(), newNews.getTargets(), publisher);
     }
     NewsUtils.broadcastEvent(NewsUtils.PUBLISH_NEWS, news.getId(), news);
     sendNotification(publisher, news, NotificationConstants.NOTIFICATION_CONTEXT.PUBLISH_IN_NEWS);
